2022/days/day03_test.go
package days

import (
	"aoc-2022/util"
	"testing"
)

func TestDay3Part1(t *testing.T) {
	exampleInp := []string{
		"vJrwpWtwJgWrhcsFMMfFFhFp",
		"jqHRNqRjqzjGDLGLrsFMfFZSrLrFZsSL",
		"PmmdzqPrVvPwwTWBwg",
		"wMqvLMZHhHMvwLHjbvcjnnSBnvTQFn",
		"ttgJtRGJQctTZtZT",
		"CrZsJsPPZsGzwwsLwLmpwMDw",
	}

	tcs := []struct {
		name string
		inp  []string
		exp  int
	}{
		{"Example", exampleInp, 157},
		{"Actual", util.ReadFileLines("../inputs/day03-1.txt"), 8401},
	}
	for _, tc := range tcs {
		t.Run(tc.name, func(t *testing.T) {
			res := day3Part1(tc.inp)

			if res != tc.exp {
				t.Fatalf("expected %d, got %d", tc.exp, res)
			}
		})
	}
}

func TestDay3Part2(t *testing.T) {
	exampleInp := []string{
		"vJrwpWtwJgWrhcsFMMfFFhFp",
		"jqHRNqRjqzjGDLGLrsFMfFZSrLrFZsSL",
		"PmmdzqPrVvPwwTWBwg",
		"wMqvLMZHhHMvwLHjbvcjnnSBnvTQFn",
		"ttgJtRGJQctTZtZT",
		"CrZsJsPPZsGzwwsLwLmpwMDw",
	}

	tcs := []struct {
		name string
		inp  []string
		exp  int
	}{
		{"Example", exampleInp, 70},
		{"Actual", util.ReadFileLines("../inputs/day03-1.txt"), 2641},
	}
	for _, tc := range tcs {
		t.Run(tc.name, func(t *testing.T) {
			res := day3Part2(tc.inp)

			if res != tc.exp {
				t.Fatalf("expected %d, got %d", tc.exp, res)
			}
		})
	}
}
